#27 feat: click outside => close dropdown 이벤트를 useDetectClickOutside 훅으로 핸들링하게 변경
